[front/components/agent_builder] - refactor: update observability components
 - Replaced ToolExecutionChart with ToolUsageChart to handle both version and step views
 - Removed ToolStepIndexChart as its functionality is merged into ToolUsageChart
 - Introduced a toggle between 'version' and 'step' modes in ToolUsageChart for improved flexibility
